pyproject.toml
[tool.poetry]
name = "Alice"
version = "1.0.0"
description = "Web-user identification is a hot research topic on the brink of sequential pattern mining and behavioral psychology."
authors = ["alexpetr0v <alexpetrov090599@gmail.com>"]
license = "None"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.23.2"
pandas = "^1.4.2"
matplotlib = "^3.5.2"
statsmodels = "^0.13.2"
scikit-learn = "^1.0.2"
xgboost = "^1.6.0"
lightgbm = "^3.3.2"
click = "^8.1.3"
dvc = {extras = ["s3"], version = "^2.11.0"}
mlflow = "^1.26.0"
python-dotenv = "^0.20.0"
pydantic = "^1.9.0"
seaborn = "^0.11.2"
catboost = "^1.0.6"
fastapi = "^0.80.0"
cachetools = "^5.2.0"
mergedeep = "^1.3.4"
uvicorn = "^0.18.3"
great-expectations = "^0.15.20"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
